[fix](regression) fix regression failure of window_funnel caused by duplicated table name
