+++ regression-test/data/nereids_p0/aggregate/window_funnel.out
@@ -11,6 +11,9 @@
 -- !window_funnel --
 2
 
+-- !window_funnel_deduplication_compat --
+4
+
 -- !window_funnel_deduplication --
 2
