--- regression-test/data/nereids_p0/sql_functions/window_functions/window_funnel.out
@@ -1,84 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !window_funnel --
-0
-
--- !window_funnel_2 --
-0
-
--- !window_funnel_3 --
-4
-
--- !window_funnel_4 --
-1	2022-03-12T19:05:04	4
-1	2022-03-12T16:15:01	3
-1	2022-03-12T13:28:02	2
-1	2022-03-12T10:41	1
-
--- !window_funnel_5 --
-1
-
--- !window_funnel_6 --
-2
-
--- !window_funnel_7 --
-0
-
--- !window_funnel_8 --
-1
-
--- !window_funnel_9 --
-1
-
--- !window_funnel_10 --
-1
-
--- !window_funnel_11 --
-1
-
--- !window_funnel_12 --
-1
-
--- !window_funnel_13 --
-1
-
--- !window_funnel_14 --
-1
-
--- !window_funnel_15 --
-1
-
--- !window_funnel_16 --
-1
-
--- !window_funnel_17 --
-1
-
--- !window_funnel_18 --
-1
-
--- !window_funnel_19 --
-1
-
--- !window_funnel_20 --
-1
-
--- !window_funnel_21 --
-1
-
--- !window_funnel_22 --
-1
-
--- !window_funnel_23 --
-1
-
--- !window_funnel_24 --
-1
-
--- !window_funnel_25 --
-1
-
--- !window_funnel_26 --
-users	13
-browser	10
-buy	1
-
