--- regression-test/data/query_p0/aggregate/window_funnel.out
@@ -1,31 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !window_funnel --
-1
-
--- !window_funnel --
-2
-
--- !window_funnel --
-1
-
--- !window_funnel --
-2
-
--- !window_funnel_deduplication_compat --
-4
-
--- !window_funnel_deduplication --
-2
-
--- !window_funnel_fixed --
-2
-
--- !window_funnel_fixed --
-1
-
--- !window_funnel_increase --
-4
-
--- !window_funnel_increase --
-2
-
